+++ R/r2.R
@@ -272,6 +272,16 @@ r2.betareg <- function(model, ...) {
 }
 
 
+#' @export
+r2.rma <- function(model, ...) {
+  s <- summary(model)
+  out <- list(R2 = s$R2 / 100)
+
+  attr(out, "model_type") <- "Meta-Analysis"
+  structure(class = "r2_generic", out)
+}
+
+
 
 #' @export
 r2.feis <- function(model, ...) {